cpufreq: activeffect: add new activeffect governor
This is the initial version of activeffect cpufreq governor, which
is a brand new governor. It is derived from interactive and ondemand
governor.

With the activeffect, the cpufreq will continuous increase when cpu
is active, and it will only decrease the cpufreq when the cpu go to
idle. That's why I call it activeffect.

It will also predict the idle time and decrease the cpu frequency if
it sleep for a long time. In this way we can decrease the cpufreq in
advance by the predictable sleep time.

There also has active boost mode, that means, if the cpu continuous
active for a specified period, then the cpu will boost its frequency
to the boost_freq, which can be set from sysfs interface.

Signed-off-by: Lianwei Wang <lianwei.wang@gmail.com>